backend/requirements.txt
boto3==1.37.13
botocore==1.37.13
Flask==3.1.0
mysql_connector_repackaged==0.3.1
PyJWT==2.10.1
PyJWT==2.10.1
pymongo==4.11.2
pytest==8.3.5
python-dotenv==1.0.1
Werkzeug==3.1.3
